% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_images.R
\name{collect_images}
\alias{collect_images}
\title{Extract array of image data from folder}
\usage{
collect_images(path, width = 210, height = 300)
}
\arguments{
\item{path}{path to folder with images}

\item{width}{resize width}

\item{height}{resize height}
}
\value{
array with first dimension as image number, and 2nd and 3rd as image dimensions
}
\description{
http://www.di.fc.ul.pt/~jpn/r/GraphicalTools/EBImage.html
}
\examples{
collect_images("data/documents_processed/b12754")
}
