% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magick_render_pdf.R
\name{magick_render_pdf}
\alias{magick_render_pdf}
\title{Render Raw Vector into PDF}
\usage{
magick_render_pdf(vector, path)
}
\arguments{
\item{vector}{vector of raw data from s3}

\item{path}{file path where object is rendered}
}
\value{
creates a pdf in the file path
}
\description{
Turn raw vector into a pdf after specifying a path
}
\examples{
s3 <- paws::s3()
content <- s3 \%>\% s3_get_document(loan_application_id = 950, attachment_id = 12227)
content$Body \%>\% magick_render_pdf(path = "document.pdf")
}
