% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magick_split_pdf.R
\name{magick_split_pdf}
\alias{magick_split_pdf}
\title{Split PDF into separate pages}
\usage{
magick_split_pdf(file_name, retained_pages, path_folder)
}
\arguments{
\item{file_name}{path to pdf}

\item{retained_pages}{pages in the pdf to capture and retain}

\item{path_folder}{folder to dump results}
}
\description{
Split PDF into separate pages
}
\examples{
magick_split_pdf(file_name = "data/documents/b9211_la9132_a17927.pdf", retained_pages = c(1, 3, 5, 7, 9, 11), path_folder = "data/documents_processed")
}
