% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_get_document.R
\name{s3_get_document}
\alias{s3_get_document}
\title{Pull Document from S3}
\usage{
s3_get_document(s3, loan_application_id, attachment_id)
}
\arguments{
\item{s3}{the s3 object created from the `paws::s3()` function}

\item{loan_application_id}{id of the loan application}

\item{attachment_id}{id of the attachment}
}
\value{
The object retrieved from the specified ids
}
\description{
Extracts document object from S3 based on loan application id and
document id
}
